wip building output

+++ resolve_templates.sh
@@ -2,6 +2,13 @@
 #
 # run helm template on chart
 #
+
+if [[ ! -d charts/impala-coordinator ]]
+then
+  echo "must be in DWX home"
+  exit 1
+fi
+
 BUILD_DIR=/tmp/chart_build
 rm -rf ${BUILD_DIR}
 mkdir ${BUILD_DIR}
@@ -13,4 +20,13 @@ cp ${BUILD_DIR}/impala-coordinator/Chart.yaml ${TMP_FILE}
 grep -v dwx-common ${TMP_FILE} >${BUILD_DIR}/impala-coordinator/Chart.yaml
 # build .tgz file
 tar -C ${BUILD_DIR} -czf /tmp/x.tar.tgz impala-coordinator
-helm3 template /tmp/x.tar.tgz
+
+for saml in true false
+do
+  echo "saml=$saml"
+done
+for proxy in true false
+do
+  echo "proxy=$proxy"
+done
+helm3 --set 'samlCallbackUrl=true,impalaEnableProxy=true'  template /tmp/x.tar.tgz > /tmp/ssss
